refactor(auth): split session.ts into types / cookies / session (#157)
Extract SessionContext, AuthEnv, SESSION_COOKIE, SESSION_TTL_SECONDS → auth/types.ts.
Extract sessionCookie, clearSessionCookie, readSessionToken → auth/cookies.ts.
session.ts retains D1 ops + middleware only. Zero behavior change; 517 tests pass.
