+++ worker/src/api/issues.ts
@@ -9,7 +9,7 @@
  */
 
 import type { Context } from "hono";
-import type { AuthEnv } from "../auth/session";
+import type { AuthEnv } from "../auth/types";
 import { resolveVisibleRepos } from "../auth/repoAccess";
 
 const ISSUE_KEY_RE = /^[A-Za-z0-9._-]+\/[A-Za-z0-9._-]+#[0-9]+$/;
