+++ worker/src/api/me.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, afterEach, vi } from "vitest";
 import { Hono } from "hono";
 import type { Env } from "../types";
 import { meRoute, logoutRoute } from "./me";
-import type { AuthEnv, SessionContext } from "../auth/session";
+import type { AuthEnv, SessionContext } from "../auth/types";
 import { requireSession } from "../auth/session";
 import type { FakeResult, FakeStmt } from "../test-utils";
 import { makeFakeStmt, makeFakeDb, captureDb, captureDbWithRows } from "../test-utils";
