+++ worker/src/api/me.ts
@@ -4,8 +4,9 @@
  */
 
 import type { Context } from "hono";
-import type { AuthEnv } from "../auth/session";
-import { readSessionToken, deleteSession, clearSessionCookie } from "../auth/session";
+import type { AuthEnv } from "../auth/types";
+import { readSessionToken, clearSessionCookie } from "../auth/cookies";
+import { deleteSession } from "../auth/session";
 
 // ---------------------------------------------------------------------------
 // GET /api/me
